重新选择界面布局方式。
